+++ cpp/hittables/hittable_list.h
