+++ cpp/hittables/moving_sphere.h
