+++ cpp/image/camera.h
