+++ cpp/image/color.h
