+++ cpp/image/image_data.h
