+++ cpp/ray_tracer.h
