+++ cpp/scenes/scenes.h
@@ -2,10 +2,22 @@
 #define WORLD_H
 
 #include "hittable_list.h"
-#include "sphere.h"
+#include "color.h"
+#include "vec3.h"
+
 #include "materials.h"
-#include "materials.cpp"
+#include "lambertian.h"
+#include "metal.h"
+#include "dielectric.h"
+
+#include "sphere.h"
+#include "moving_sphere.h"
 #include "camera.h"
+
+#include "texture.h"
+#include "solid_color.h"
+#include "checker_texture.h"
+#include "noise_texture.h"
 
 HittableList get_world()
 {
@@ -23,45 +35,9 @@ HittableList get_world()
     return world;
 }
 
-HittableList get_depth_of_field_example(Camera &camera)
-{
-    Vec3 look_from(3, 3, 2);
-    Vec3 look_at(0, 0, -1);
-    double dist_to_focus = (look_from - look_at).length();
-    double aperture = 2.0;
-
-    camera = Camera(
-        20,            // vertical field of view
-        16.0 / 9.0,    // aspect ratio
-        aperture,      // aperture
-        dist_to_focus, // focus distance
-        look_from,     // origin
-        look_at        // look at
-    );
-
-    Material *material_ground = new Lambertian(Color(0.8, 0.8, 0.0));
-    Material *material_blue = new Lambertian(Color(0.1, 0.2, 0.5));
-    Material *material_glass = new Dielectric(1.5);
-    Material *material_metal = new Metal(Color(0.8, 0.6, 0.2), 0.0);
-
-    auto sphere_ground = std::make_shared<Sphere>(Vec3(0.0, -100.5, -1.0), 100.0, material_ground);
-    auto sphere_glass = std::make_shared<Sphere>(Vec3(-1.0, 0.0, -1.0), 0.5, material_glass);
-    auto sphere_glass_air = std::make_shared<Sphere>(Vec3(-1.0, 0.0, -1.0), -0.45, material_glass);
-    auto sphere_blue = std::make_shared<Sphere>(Vec3(0.0, 0.0, -1.0), 0.5, material_blue);
-    auto sphere_metal = std::make_shared<Sphere>(Vec3(1.0, 0.0, -1.0), 0.5, material_metal);
-
-    HittableList world = HittableList();
-    world.add(sphere_ground);
-    world.add(sphere_glass);
-    world.add(sphere_glass_air);
-    world.add(sphere_blue);
-    world.add(sphere_metal);
-    return world;
-}
-
 HittableList get_random_scene(Camera &camera)
 {
-    camera = Camera(
+    camera.setUp(
         20,             // vertical field of view
         16.0 / 9.0,     // aspect ratio
         0.1,            // aperture
@@ -130,6 +106,18 @@ HittableList get_random_scene(Camera &camera)
 HittableList get_random_scene_moving(Camera &camera)
 {
 
+    camera.setUp(
+        20,             // vertical field of view
+        16.0 / 9.0,     // aspect ratio
+        0.1,            // aperture
+        10.0,           // focus distance
+        Vec3(13, 2, 3), // origin
+        Vec3(0, 0, 0),  // look at
+        0,              // time 0
+        1               // time 1
+    );
+    std::cerr << "Moving scenario" << std::endl;
+
     HittableList world;
 
     Material *ground_material = new Lambertian(Color(0.5, 0.5, 0.5));
@@ -193,7 +181,20 @@ HittableList get_random_scene_moving(Camera &camera)
 HittableList get_random_scene_checker(Camera &camera)
 {
 
+    camera.setUp(
+        20,             // vertical field of view
+        16.0 / 9.0,     // aspect ratio
+        0.1,            // aperture
+        10.0,           // focus distance
+        Vec3(13, 2, 3), // origin
+        Vec3(0, 0, 0),  // look at
+        0,              // time 0
+        1               // time 1
+    );
+
     HittableList world;
+
+    std::cerr << "Checker scenario" << std::endl;
 
     Texture *texture_a = new SolidColor(Color(0.2, 0.3, 0.1));
     Texture *texture_b = new SolidColor(Color(0.9, 0.9, 0.9));
